fix(ci): remove duplicate-code ratchet violations
Why\nKeep the full verification gate green after the current mainline changes introduced duplicate-code ratchet failures.\n\nWhat\nExtract the shared git command helper, reuse the canonical Temporal follow-up schema, and refresh the jscpd baseline for the resulting smaller clone set.\n\nVerification\n- bun run jscpd\n- bunx turbo run typecheck test lint --filter=@shepherdjerred/birmel --filter=@shepherdjerred/temporal
