+++ .jscpd-baseline.json
@@ -120,10 +120,6 @@
       "clones": 1,
       "tokens": 86
     },
-    "packages/birmel/src/editor/github-pr.ts|packages/birmel/src/editor/repo-clone.ts": {
-      "clones": 1,
-      "tokens": 129
-    },
     "packages/birmel/src/health/server.ts|packages/starlight-karma-bot/src/server/index.ts": {
       "clones": 1,
       "tokens": 93
@@ -1829,8 +1825,8 @@
       "tokens": 141
     },
     "packages/temporal/src/shared/agent-task.ts|packages/temporal/src/shared/agent-task.ts": {
-      "clones": 2,
-      "tokens": 349
+      "clones": 1,
+      "tokens": 143
     },
     "packages/temporal/src/shared/llm-catalog-alert.test.ts|packages/temporal/src/shared/llm-catalog-alert.test.ts": {
       "clones": 1,
@@ -1904,5 +1900,5 @@
       "clones": 3,
       "tokens": 238
     }
-  }
+  },
 }
