+++ packages/birmel/src/editor/repo-clone.ts
@@ -1,6 +1,7 @@
 import { rm, mkdir } from "node:fs/promises";
 import path from "node:path";
 import { loggers } from "@shepherdjerred/birmel/utils/logger.ts";
+import { runGitCommand } from "./git-command.ts";
 
 const logger = loggers.editor.child("repo-clone");
 
@@ -70,26 +71,4 @@ export async function cleanupClone(clonePath: string): Promise<void> {
       path: clonePath,
     });
   }
-}
-
-/**
- * Run a git command in a directory
- */
-async function runGitCommand(cwd: string, args: string[]): Promise<string> {
-  const proc = Bun.spawn(["git", ...args], {
-    cwd,
-    stdout: "pipe",
-    stderr: "pipe",
-  });
-
-  const [stdout, stderr, exitCode] = await Promise.all([
-    new Response(proc.stdout).text(),
-    new Response(proc.stderr).text(),
-    proc.exited,
-  ]);
-
-  if (exitCode === 0) {
-    return stdout.trim();
-  }
-  throw new Error(`git ${args.join(" ")} failed: ${stderr}`);
 }
